Implemented Locale in app. If Locale is in America, price is accepted and filtered in USD, if Locale is in a Euro Currency Confining country, price is accepted as Euro and filtered as Euro.
If Locale is outside USA, Dates are taken and filtered in dd/MM//YY format,
Otherwise if Locale is in USA, Dates are taken and filtered in MM/dd//YY format.

Also, Login Activity has been update with a "Sign in as Guest" button, through which the user can visit the all listings and map fragments to view the listings and even click on them to view a listing in detail, but they cannot add a listing or update a listing until they are logged in.